requirements.txt
Django==3.0.8
djangorestframework==3.11.0
Faker==4.1.1
celery==4.4.6
redis==3.5.3
psycopg2==2.8.5
